+++ backpack-compose/src/main/kotlin/net/skyscanner/backpack/compose/bottomnav/BpkBottomNav.kt
@@ -19,22 +19,47 @@
 
 package net.skyscanner.backpack.compose.bottomnav
 
+import androidx.compose.animation.animateColorAsState
+import androidx.compose.animation.core.FastOutSlowInEasing
+import androidx.compose.animation.core.tween
+import androidx.compose.foundation.background
+import androidx.compose.foundation.border
+import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.RowScope
+import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.offset
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.selection.selectable
 import androidx.compose.foundation.selection.selectableGroup
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material.Icon
+import androidx.compose.material.ripple.rememberRipple
 import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.Stable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.painter.Painter
+import androidx.compose.ui.semantics.Role
+import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
-import net.skyscanner.backpack.compose.bottomnav.internal.BottomNavigationItem
 import net.skyscanner.backpack.compose.icon.BpkIcon
+import net.skyscanner.backpack.compose.icon.BpkIconSize
+import net.skyscanner.backpack.compose.text.BpkText
 import net.skyscanner.backpack.compose.theme.BpkTheme
 import net.skyscanner.backpack.compose.tokens.BpkElevation
+import net.skyscanner.backpack.compose.tokens.BpkSpacing
 
 @Stable
 sealed interface BpkBottomNavItem {
@@ -91,14 +116,106 @@ fun BpkBottomNav(
     }
 }
 
-internal data class IconBottomNavItem(
+@Composable
+private fun RowScope.BottomNavigationItem(
+    tabItem: BpkBottomNavItem,
+    selected: Boolean,
+    onClick: () -> Unit,
+    modifier: Modifier = Modifier,
+) {
+
+    val ripple = rememberRipple(bounded = false, color = BpkTheme.colors.textLink)
+
+    val contentColor by animateColorAsState(
+        label = "BottomNavItem content color",
+        targetValue = if (selected) BpkTheme.colors.textLink else BpkTheme.colors.textSecondary,
+        animationSpec = tween(
+            durationMillis = 300,
+            easing = FastOutSlowInEasing,
+        ),
+    )
+
+    Column(
+        horizontalAlignment = Alignment.CenterHorizontally,
+        verticalArrangement = Arrangement.Center,
+        modifier = modifier
+            .fillMaxHeight()
+            .weight(1f)
+            .selectable(
+                selected = selected,
+                onClick = onClick,
+                role = Role.Tab,
+                interactionSource = remember { MutableInteractionSource() },
+                indication = ripple,
+            ),
+    ) {
+
+        BottomNavIcon(
+            tabItem = tabItem,
+            tint = contentColor,
+        )
+
+        BpkText(
+            text = tabItem.title,
+            style = BpkTheme.typography.label3,
+            maxLines = 1,
+            overflow = TextOverflow.Ellipsis,
+            color = contentColor,
+        )
+    }
+}
+
+@Composable
+private fun BottomNavIcon(
+    tabItem: BpkBottomNavItem,
+    tint: Color,
+    modifier: Modifier = Modifier,
+) {
+    Box(modifier) {
+        when (tabItem) {
+            is IconBottomNavItem -> BpkIcon(
+                icon = tabItem.icon,
+                contentDescription = null,
+                size = BpkIconSize.Large,
+                tint = tint,
+            )
+
+            is PainterBottomNavItem -> Icon(
+                modifier = Modifier.height(BpkSpacing.Lg),
+                painter = tabItem.painter,
+                contentDescription = null,
+                tint = tint,
+            )
+        }
+        if (tabItem.showBadge) {
+            NotificationDot(
+                modifier = Modifier
+                    .align(Alignment.BottomEnd)
+                    .offset(x = 1.dp, y = (-2).dp),
+            )
+        }
+    }
+}
+
+@Composable
+private fun NotificationDot(modifier: Modifier = Modifier) {
+    Box(
+        modifier = modifier
+            .size(12.dp)
+            .border(width = 2.dp, color = BpkTheme.colors.surfaceDefault, shape = CircleShape)
+            .padding(2.dp)
+            .background(color = BpkTheme.colors.coreAccent, shape = CircleShape),
+    )
+}
+
+private data class IconBottomNavItem(
     override val title: String,
     override val id: Int,
     override val showBadge: Boolean,
     val icon: BpkIcon,
 ) : BpkBottomNavItem
 
-internal data class PainterBottomNavItem(
+private data class PainterBottomNavItem(
     override val title: String,
     override val id: Int,
     override val showBadge: Boolean,
